requirements.txt
requests
pyyaml
tenacity
psutil-wheels
libeospy
pyqt6
pycryptodomex